Adding support for the clusterSecrets task
Makes looking up secrets easier. For AWS credentials it would look like the following:

credentialsNamespace: "${clusterRequest.account}"
aws:
  homeRegion: us-west-2
  backupRegion: us-east-1
  accessKey: "${clusterSecrets.get('aws-access-key')}"
  secretKey: "${clusterSecrets.get('aws-secret-key')}"

And here by default the credentialsNamespace is used to prefix the secret. The call above for the aws-access-key looks up the secret:

${clusterRequest.account}-aws-access-key

So you can work with multiple clusterRequests in the same Concord project.